+++ samples/webpack-app-hello-pug/src/main.js
@@ -1,4 +1,4 @@
-import { app } from './app/app';
+import { app } from 'App/app';
 
 window.addEventListener('load', function (event) {
   console.log('Start app!');