+++ test/cases/javascript-assets-html/src/index.js
@@ -1,0 +1,2 @@
+const html = require('Template/widget.pug');
+console.log(html);